Add closePopUp file
